requirements.txt
flask
twilio
python-dotenv
easyocr
gtts
fpdf2
pillow
numpy
streamlit
pandas
anthropic
openpyxl
pyttsx3
